refactor utils and constants
