+++ olive/systems/utils.py
@@ -2,8 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-import argparse
-import json
 import logging
 import shutil
 import tempfile
@@ -11,57 +9,14 @@ from functools import lru_cache
 from pathlib import Path
 
 from olive.common.utils import hash_dir
+from olive.hardware.constants import PROVIDER_DOCKERFILE_MAPPING, PROVIDER_PACKAGE_MAPPING
 from olive.systems.common import SystemType
 
 logger = logging.getLogger(__name__)
 
 
-def parse_common_args(raw_args):
-    """Parse common args."""
-    parser = argparse.ArgumentParser("Olive common args")
-
-    # model args
-    parser.add_argument("--model_config", type=str, help="Path to model config json", required=True)
-
-    # pipeline output arg
-    parser.add_argument("--pipeline_output", type=str, help="pipeline output path", required=True)
-
-    return parser.parse_known_args(raw_args)
-
-
-def parse_model_resources(model_resource_names, raw_args):
-    """Parse model resources. These are the resources that were uploaded with the job."""
-    parser = argparse.ArgumentParser("Model resources")
-
-    # parse model resources
-    for resource_name in model_resource_names:
-        # will keep as required since only uploaded resources should be in this list
-        parser.add_argument(f"--model_{resource_name}", type=str, required=True)
-
-    return parser.parse_known_args(raw_args)
-
-
-def get_common_args(raw_args):
-    """Return the model_config.
-
-    The return value includes json with the model resource paths filled in, the pipeline output path, and any
-    extra args that were not parsed.
-    """
-    common_args, extra_args = parse_common_args(raw_args)
-
-    # load model json
-    with open(common_args.model_config) as f:  # noqa: PTH123
-        model_json = json.load(f)
-    # model json has a list of model resource names
-    model_resource_names = model_json.pop("resource_names")
-    model_resource_args, extra_args = parse_model_resources(model_resource_names, extra_args)
-
-    for key, value in vars(model_resource_args).items():
-        # remove the model_ prefix, the 1 is to only replace the first occurrence
-        normalized_key = key.replace("model_", "", 1)
-        model_json["config"][normalized_key] = value
-
-    return model_json, common_args.pipeline_output, extra_args
+def get_package_name_from_ep(execution_provider):
+    return PROVIDER_PACKAGE_MAPPING.get(execution_provider, ("onnxruntime", "ort-nightly"))
 
 
 @lru_cache(maxsize=8)
@@ -71,12 +26,6 @@ def create_new_system_with_cache(system_config, accelerator):
 
 def create_new_system(system_config, accelerator):
     # pylint: disable=consider-using-with
-    provider_dockerfile_mapping = {
-        "CPUExecutionProvider": "Dockerfile.cpu",
-        "CUDAExecutionProvider": "Dockerfile.gpu",
-        "TensorrtExecutionProvider": "Dockerfile.gpu",
-        "OpenVINOExecutionProvider": "Dockerfile.openvino",
-    }
 
     # create a new system with the same type as the origin system
     if system_config.type == SystemType.Local:
@@ -117,7 +66,7 @@ def create_new_system(system_config, accelerator):
     elif system_config.type == SystemType.Docker:
         from olive.systems.docker import DockerSystem
 
-        dockerfile = provider_dockerfile_mapping.get(accelerator.execution_provider, "Dockerfile.cpu")
+        dockerfile = PROVIDER_DOCKERFILE_MAPPING.get(accelerator.execution_provider, "Dockerfile.cpu")
         # TODO(myguo): create a temp dir for the build context
         new_system = DockerSystem(
             local_docker_config={
@@ -136,7 +85,7 @@ def create_new_system(system_config, accelerator):
     elif system_config.type == SystemType.AzureML:
         from olive.systems.azureml import AzureMLSystem
 
-        dockerfile = provider_dockerfile_mapping.get(accelerator.execution_provider, "Dockerfile.cpu")
+        dockerfile = PROVIDER_DOCKERFILE_MAPPING.get(accelerator.execution_provider, "Dockerfile.cpu")
         temp_dir = tempfile.TemporaryDirectory()  # pylint: disable=consider-using-with
         build_context_path = Path(temp_dir.name)
         shutil.copy2(str(Path(__file__).parent / "docker" / dockerfile), build_context_path)
